pyproject.toml
[project]
name = "motif"
version = "0.1.0"
description = "A prompt algebra for LLM orchestration. The code reads as its own specification."
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40",
    "python-dotenv>=1.2.2",
]

[project.optional-dependencies]
display = ["rich>=13"]
dev = ["pytest", "hypothesis"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/motif"]
